Update demo examples to include a video with chapters, with descriptions, and using a container when entering fullscreen
