+++ demo-laerdal/index.html
@@ -14,15 +14,29 @@
       <main>
         <div id="container">
           <!-- Video -->
-          <h2>Video</h2>
+          <h2>1. Video</h2>
           <video id="video-player" controls crossorigin playsinline></video>
 
+          <h2>2. Video with Descriptions</h2>
+          <video id="video-player-with-descriptions" controls crossorigin playsinline></video>
+
+          <h2>3. Video with Chapters</h2>
+          <video id="video-player-with-chapters" controls crossorigin playsinline></video>
+
+          <h2>4. Video with Fullscreen Container</h2>
+          <div class="container">
+            <p>Something in the container</p>
+            <button>Button to tab</button>
+            <a href="#">Link to tab</a>
+            <video id="video-player-with-fullscreen-container" controls crossorigin playsinline></video>
+          </div>
+
           <!-- Audio -->
-          <h2>Audio</h2>
+          <h2>5. Audio</h2>
           <audio id="audio-player" controls crossorigin></audio>
 
           <!-- Youtube -->
-          <h2>Youtube</h2>
+          <h2>6. Youtube</h2>
           <div class="plyr__video-embed" id="youtube-player">
             <iframe
               src="https://www.youtube.com/embed/bTqVqk7FSmY?origin=https://plyr.io&amp;iv_load_policy=3&amp;modestbranding=1&amp;playsinline=1&amp;showinfo=0&amp;rel=0&amp;enablejsapi=1"
@@ -33,7 +47,7 @@
           </div>
 
           <!-- Vimeo -->
-          <h2>Vimeo</h2>
+          <h2>7. Vimeo</h2>
           <div class="plyr__video-embed" id="vimeo-player">
             <iframe
               src="https://player.vimeo.com/video/76979871?loop=false&amp;byline=false&amp;portrait=false&amp;title=false&amp;speed=true&amp;transparent=0&amp;gesture=media"
